return to expression code
