+++ packages/perseus/src/widgets/expression.tsx
@@ -21,7 +21,12 @@ import type {
     PerseusExpressionWidgetOptions,
     PerseusExpressionAnswerForm,
 } from "../perseus-types";
-import type {PerseusScore, WidgetExports, WidgetProps} from "../types";
+import type {
+    APIOptions,
+    PerseusScore,
+    WidgetExports,
+    WidgetProps,
+} from "../types";
 import type {Keys as Key, KeypadConfiguration} from "@khanacademy/math-input";
 import type {PropsFor} from "@khanacademy/wonder-blocks-core";
 
@@ -50,6 +55,20 @@ const insertBraces = (value) => {
     //
     // TODO(alex): Properly hack MathQuill to always use explicit braces.
     return value.replace(/([_^])([^{])/g, "$1{$2}");
+};
+
+const deriveKeypadVersion = (apiOptions: APIOptions) => {
+    // We can derive which version of the keypad is in use. This is
+    // a bit tricky, but this code will be relatively short-lived
+    // as we coalesce onto the new, v2 Keypad, at which point we
+    // can remove this `virtualKeypadVersion` field entirely.
+    return apiOptions.nativeKeypadProxy != null
+        ? "REACT_NATIVE_KEYPAD"
+        : apiOptions.customKeypad === true
+        ? apiOptions.useV2Keypad === true
+            ? "MATH_INPUT_KEYPAD_V2"
+            : "MATH_INPUT_KEYPAD_V1"
+        : "PERSEUS_MATH_INPUT";
 };
 
 type Rubric = PerseusExpressionWidgetOptions;
@@ -350,14 +369,32 @@ export class Expression extends React.Component<Props, ExpressionState> {
     };
 
     simpleValidate: (
-        rubric: Rubric,
+        rubric: Rubric & {scoring?: boolean},
         onInputError: OnInputErrorFunctionType,
-    ) => PerseusScore = (rubric, onInputError) =>
-        Expression.validate(
+    ) => PerseusScore = ({scoring, ...rubric}, onInputError) => {
+        const score = Expression.validate(
             this.getUserInput(),
             rubric,
             onInputError || function () {},
         );
+
+        // "scoring" is a flag that indicates when we are checking answers.
+        // otherwise, we may just be checking validity after changes.
+        if (scoring && score.type !== "invalid") {
+            this.props.analytics?.onAnalyticsEvent({
+                type: "perseus:expression-evaluated",
+                payload: {
+                    result:
+                        score.earned === score.total ? "correct" : "incorrect",
+                    virtualKeypadVersion: deriveKeypadVersion(
+                        this.props.apiOptions,
+                    ),
+                },
+            });
+        }
+
+        return score;
+    };
 
     getUserInput: () => string = () => {
         return Expression.getUserInputFromProps(this.props);
